requirements.txt
simple-salesforce==1.12.6
fillpdf==0.7.3
python-dotenv==1.0.1
click==8.1.7
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
